+++ src/main/java/com/sinwn/capsule/mapper/UserEntityMapper.java
@@ -6,13 +6,13 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface UserEntityMapper {
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(UserEntity record);
 
     int insertSelective(UserEntity record);
 
-    UserEntity selectByPrimaryKey(Long id);
+    UserEntity selectByPrimaryKey(Integer id);
 
     int updateByPrimaryKeySelective(UserEntity record);
 
@@ -22,5 +22,5 @@ public interface UserEntityMapper {
 
     UserEntity selectByUsername(String username);
 
-    UserEntity selectByUserId(long userId);
+    UserEntity selectByUserId(Integer userId);
 }
